+++ about.html
@@ -7,10 +7,9 @@
     <link rel="stylesheet" href="css/reset.css">
     <link rel="stylesheet" href="css/grid.css">
     <link rel="stylesheet" href="css/main.css">
-    <script defer src="js/main.js"></script>
-    <script defer src="js/about.js"></script>
+    <script type="module" defer src="js/main.js"></script>
 </head>
-<body>
+<body data-page="about">
     <h1 class="hidden">About Delilah Singer</h1>
     <header class="full-width-grid-con">
         <nav class="grid-con top-nav">
