+++ admin/addthis.txt
@@ -1,0 +1,2 @@
+FOREIGN KEY (project_id) REFERENCES projects(id) 
+    ON DELETE CASCADE
