+++ contact.html
@@ -7,9 +7,9 @@
     <link rel="stylesheet" href="css/reset.css">
     <link rel="stylesheet" href="css/grid.css">
     <link rel="stylesheet" href="css/main.css">
-    <script defer src="js/main.js"></script>
+    <script type="module" defer src="js/main.js"></script>
 </head>
-<body>
+<body data-page="contact">
     <h1 class="hidden">Contact</h1>
     <header class="full-width-grid-con">
         <nav class="grid-con top-nav">
