+++ demo-api.html
@@ -7,10 +7,9 @@
     <link rel="stylesheet" href="css/reset.css">
     <link rel="stylesheet" href="css/grid.css">
     <link rel="stylesheet" href="css/main.css">
-    <script defer src="js/genshin-api.js"></script>
-    <script defer src="js/main.js"></script>
+    <script type="module" defer src="js/main.js"></script>
 </head>
-<body>
+<body data-page="genshin">
     <h1 class="hidden">API Demo</h1>
     <header class="full-width-grid-con">
         <nav class="grid-con top-nav">
