+++ index.html
@@ -13,10 +13,10 @@
     <script defer src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.12.2/ScrollToPlugin.min.js"></script>
     <script defer src="https://kit.fontawesome.com/a7c4f622e4.js" crossorigin="anonymous"></script>
     <script defer src="js/video.js"></script>
-    <script defer src="js/index.js"></script>
+    <script type="module" defer src="js/main.js"></script>
     
 </head>
-<body>
+<body data-page="index">
     <h1 class="hidden">Delilah Singer's Portfolio</h1>
     <header class="full-width-grid-con">
         <nav class="grid-con top-nav">
